% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_map.R
\name{bin_map}
\alias{bin_map}
\title{Create a binary map}
\usage{
bin_map(x, threshold)
}
\arguments{
\item{x}{A RasterLayer object}

\item{threshold}{}
}
\value{
A RasterLayer object
}
\description{
Create a binary map
}
