% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_map.R
\name{rescale_map}
\alias{rescale_map}
\title{Rescale raster values to 0-1}
\usage{
rescale_map(x)
}
\arguments{
\item{x}{The SpatRaster object to be rescaled}
}
\value{
A SpatRaster object
}
\description{
Rescale raster values to 0-1
}
