% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_map.R
\name{source_map}
\alias{source_map}
\title{Create a source map, where all values other than focal value become NA}
\usage{
source_map(x, value = 1)
}
\arguments{
\item{x}{A RasterLayer object}

\item{value}{The cell value to create source surface from, default is 1}
}
\value{
A RasterLayer object
}
\description{
Create a source map, where all values other than focal value become NA
}
