% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_map.R
\name{bin_map}
\alias{bin_map}
\title{Create a binary map}
\usage{
bin_map(x, threshold)
}
\arguments{
\item{x}{A SpatRaster object}

\item{threshold}{A value (or range of values) to use as suitability thresholds}
}
\value{
A SpatRaster object
}
\description{
This function creates a binary map of suitable/unsuitable habitat based on a user-specified threshold value
}
