% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_priority.R
\name{patch_priority}
\alias{patch_priority}
\title{Calculate habitat patch priority}
\usage{
patch_priority(suit, suit_bin)
}
\arguments{
\item{suit}{A RasterLayer object of habitat suitability values, values range from 0-1}

\item{suit_bin}{A RasterLayer object of binary suitable (1) and unsuitable (0) habitat created using \code{\link[habicon]{bin_map}}}
}
\value{
A RasterLayer object
}
\description{
Calculate habitat patch priority
}
